requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.3
Flask-Login==0.2.7
Flask-Bootstrap
Flask-Admin==1.1.0
Flask-Testing
coverage
sphinx
schedule

tweepy
